feat: infra 未学習のカード一覧取得

+++ app/infra/db/client.go
@@ -20,7 +20,7 @@ type IClient interface {
 	CreateCard(ctx context.Context, card *domain.Card) (*domain.Card, error)
 	GetCard(ctx context.Context, id int64) (*domain.Card, error)
 	GetCards(ctx context.Context, where *domain.CardsWhere, limit, offset int) ([]*domain.Card, error)
-	GetPendingCards(ctx context.Context, deckID int, to time.Time, limit, offset int) ([]*domain.Card, error)
+	GetPendingCards(ctx context.Context, deckID int64, to time.Time, limit, offset int) ([]*domain.Card, error)
 	UpdateCardByID(ctx context.Context, id int64, input *domain.UpdateCardInput) (*domain.Card, error)
 	CountCards(ctx context.Context, where *domain.CardsWhere) (int, error)
